# apps/storefront/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Gorumin BFF API",
        "version": "1.0.0",
        "description": "Backend-for-frontend del storefront de Gorumin. Proxy tipado hacia Medusa con CORS, rate limiting y sesi\u00f3n de cliente en cookie httpOnly."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local"
        },
        {
            "url": "https://gorumin.com/api",
            "description": "Producci\u00f3n"
        }
    ],
...
    "components": {
        "parameters": {
...
            "region": {
                "name": "region",
                "in": "query",
                "description": "C\u00f3digo ISO2 del pa\u00eds (default: co)",
                "schema": {
                    "type": "string",
                    "default": "co"
                }
            }
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "_gorumin_jwt"
            }
        }
    },
    "paths": {
...
        "/products/{handle}": {
            "get": {
                "tags": [
                    "catalog"
                ],
                "summary": "Detalle de producto por handle",
                "parameters": [
                    {
                        "name": "handle",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/region"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/cart": {
            "get": {
                "tags": [
                    "cart"
                ],
                "summary": "Obtener carrito actual (cookie o ?id=)",
                "responses": {
                    "200": {
                        "description": "Carrito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "cart"
                ],
                "summary": "Crear carrito",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "region": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carrito creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cart/{id}": {
            "post": {
                "tags": [
                    "cart"
                ],
                "summary": "Actualizar carrito (email, direcciones, etc.)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carrito actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cart/{id}/line-items": {
            "post": {
                "tags": [
                    "cart"
                ],
                "summary": "Agregar l\u00ednea al carrito",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "variant_id"
                                ],
                                "properties": {
                                    "variant_id": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "default": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Carrito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cart/{id}/line-items/{line}": {
            "post": {
                "tags": [
                    "cart"
                ],
                "summary": "Actualizar cantidad de una l\u00ednea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "line",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carrito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "cart"
                ],
                "summary": "Eliminar una l\u00ednea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "line",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carrito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}